requirements.txt
pandas
numpy
scikit-learn
prophet
dash
plotly
requests
sqlalchemy
psycopg2-binary
python-dotenv
apscheduler
